requirements.txt
selenium ~= 4.4.3
chromedriver-binary-auto ~= 0.2.0